dev: support for running container test targets
Provides a tooling-agnostic way of running tests. It's just an initial
starting point and will hopefully be replaced by something much better.

The pattern is simple, every service (aka lens) has a "test" entry
generated in the giant docker-compose yaml we create. To run it, use
a command like this:

./dev.sh docker-compose-build substrate-test

This builds the "test" target in images/substrate/Dockerfile

Which, in this case, runs the appropriate `go test` command

Also remove unused code that was creating test failures.