feat: enable live editing for substrate ui
Switch to a top-level "live_edit" field for all services. This is
a simpler convention than our previous approach.

Remove the now vestigial EXTERNAL_UI_HANDLER override for substrate

Also add an optimization to system-reload for the common case of
"not every image has changed".

Add a check that the systemd units generated for containers are valid.