feat: substrate gateway works properly with spaces
Substrate provides a mechanism for containers to request persistent
storage. This mechanism is called a "space".

Spaces can be created, checkpointed, and forked.

By default checkpoints and forks are done in the naive and O(N) way,
but once we settle on a checkpointable filesystem for persistent data,
they will be O(1).

To see these in action, visit /gotty.

You should notice you are redirected to another URL that includes a
bit of text that looks like "data=sp-...". This means substrate has
allocated a space and aliased it to the "data" slot in the container.

You can visit this new URL in multiple browsers and see that each
connection is to the same container instance.

Anything stored in /spaces/data/tree will survive that container's
restart.

A future change will add support to bridge2 for saving to a space.

While we're rearranging how things are spawned, drop ExtraEnvironment
in favor of a simpler approach.

Also fix our gotty Dockerfile to be a bit better behaved