refactor: move /bb out to be its own service
Previously, the /bb HTTP endpoint ran within substrate. It would load
the same defs that substrate did and find all the matches for an
incoming request, forwarding the incoming request to all of them in
parallel. This meant spawning containers and reverse proxying traffic
to them. It would then return the first valid response to the original
HTTP request.

Now, /bb is a service just like any other service in substrate. It is
lazily spawned on first request, it also no longer has direct access to
spawning containers or the defs that substrate uses internally.

To support this, there's a new endpoint inside of substrate at
/substrate/v1/defs. This endpoint is formatted as a text/event-stream
and emits data every time any substrate def changes. That data contains
a large JSON object of all the .cue files that define the
substrate defs.

When it boots, /bb subscribes to /substrate/v1/defs, evaluates them
on each update, and uses that for all new incoming requests. Note that
it waits for the first def announcement before it serves any requests.

When it receives a request, /bb looks at the current set of defs it
has and finds all services that might serve matching requests. It
issues requests back to substrate for all of them. The first valid
response is returned to the requester. This reduces the codepaths
inside of substrate and both substrate and bb easier to reason about.

This approach opens the door to three future changes. One is a
simplification and the other two are big feature upgrades for
the system.

The simplification that becomes possible is moving "system resource
information" into their own services (like how many CUDA cards? and
how much CUDA memory do they have? how much RAM do we have?). Moving
these out to be standalone makes them both easier to understand and
makes them easier to run in real-time. This would allow substrate to
make better resource allocations decisions, including which models to
unload, what cards to run them on, and whether or not it should run
small or large versions of various models.

The first feature upgrade is providing bb with its own storage. Since
it's just another service it can store files in a space like any other
service. With persistent storage we can allow users to add new matching
rules (analogous to prolog facts) over time. These rules might include
things like "triggers" or "semantic reminders". Another use can be
user-authored "meta strategies".

The second feature upgrade is allowing separate instances of bb to run
in a sort of hierarchy. This could be different "work layers", like one
specific to a person, one specific to a physical room, one specific to
a project. It can also be different devices... like if you're running
multiple substrate boxes and want some of them to do vision tasks,
others to do retrieval tasks, others still might drive
real-world activities.