bridge2: only redirect requests for `/`
The http ServeMux pattern `/` any requests that
aren't already assigned, so this would redirect
other unknown paths to `/sessions` and create a
new session. So on page loads, extra requests
like `/favicon.ico` would trigger a new session
in the background.

This only redirects when the path is just `/`.

Additionally we could have the `GET /session`
index not create a session, and leave that to a JS
POST call on load, but refining the redirect is
good enough for the moment.

Fixes #123