refactor: rename lens => service throughout
The initial term "lens" was used in a previous iteration of substrate
that was much more "space" centric. Using multiple "lenses" to interact
with a "space" was a useful metaphor. However, people seemed to have
trouble remembering the word "lens" and/or what it represented.

We *previously* used the term "service" to refer to always-running
processes that defined the core system. These are now called "daemons".

In this latest iteration of substrate we had many "lenses" that didn't
actually use spaces at all. We also found ways to switch many of the
core "services" into "lenses".

Given our new architecture and all that we've learned, renaming
"service" -> "daemon" and "lens" -> "service" is a good trade.

We do this rename now just before we start writing core data structures
to disk, to avoid carrying these legacy terms into our data formats.