feat: record service spawns and show their history
Recording spawns means we can use the history of spawns to "go back to"
previously allocated spaces.

This makes it much easier to use stateful experiences like bridge.

Refactor all the db reading and writing logic to be in a separate
package. Its implementation should be fairly uninteresting for most
people looking at substrate and doesn't need to be in the main package.

Re-enable screenshot, visualizer, files services so the UI is a bit
better behaved.

Continue to reduce our use of the "activity" metaphor in favor of the
"service" metaphor. The concept of an "activity" might return, but for
now we can get quite far with "services" alone. For reference, a service
is a running container serving HTTP requests and an activity is a
particular URL path for a given service.

The idea was to provide a way to discover which URLs at which services
could be used in a given situation. This is of course still a valuable
abstraction, but it wasn't really being used to solve any critical
problems and it was getting in the way.

Introduce the idea of an "ephemeral" service. This is a service whose
spawn should not be considered noteworthy from a history tracking
perspective. This is probably not quite the right word, but it's at
least close. Spawn events for ephemeral services are not written
to disk.