bridge2: don't start new session on index
It's not always appropriate to start a new session
when loading the index, so move new session
creation to a new path. Adds a button to the index
to start a session, or the user can select a
previous session from the sidebar.

Some other small UI updates to highlight the
active session in the sidebar, and show the
session time instead of "session1" header.

Also adds a timeout for refreshing the list of
sessions when there is not active session, or
there are no new events on the current session.

Fixes #161