feat: an initial primer
Primers are an initial step towards a new and experimental form of
media. They are multiple things at once:
- a "simulator" for intricate substrate-to-substrate interactions
- a storytelling platform, to explain multiuser-based scenarios
- a integration test platform for substrate functionality
- media production tool to showcase substrate capabilities and benefits

This implementation is based on the @observablehq/framework npm package.

They also make use of chromestage and substrate commands to directly
embed and automate various interactions.

A casual user sees a primer as a set of templated scenes that show one
or more characters in a specific situation interacting with each other
and one or more devices. We present different camera angles for each
scene according to an event sequence described that primer's
source code. We also allow forward and backward "scrubbing" of
events to make it easier explore and dissect what is happening at any
given point.

Real screen contents are composited into the scene by finding and
replacing SVG shapes on-the-fly. To keep this simple, we assign each
screen a solid and unique fill color and refer to it using color.

Screen contents are backed by real, dynamic and optionally interactive
streams of virtual framebuffers are actually running the code in
question.

Future work will:
- make these screen streams easy to record and replay
- allow screens to export audio to each other and the user
- allow "export" of a primer as a traditional video
- allow primers to "self edit" during their execution
- allow users to "fork" the execution and "pop out" individual screens
- fully nest all state of substrate itself, including fs and bb defs