chore: remove cruft from images/files
We previously delegated jupyter notebook previews to the files service.
This may or may not be what we choose to do in the future. Remove this
and other bits of cruft from the previous implementation.

Replace the (increasingly inaccurate) images/files README.md with a
reference to the original upstream project.