bridge2: assistant agent (#182)
Adds a basic assistant agent. Right now configuration is mostly hard-coded to include the "bridge" assistant. More could be added to the map, but we can address runtime configuration in #190.
It looks for the assistant name anywhere in the message and passes just the current message to the configured URL using the OpenAI API.

Fixes #64