chore: rename "spawn" field to "instances[string]"
This name better communicates the anticipated "cardinality" of what we
spawn when we route traffic to a specific service instance. It also
simplifies how we can use exports from a service (like /nvml for GPU
memory or /sigar for total available RAM) to inform how we can spawn
other services.