chore: optionally pre-resolve image references
In a future change we'll support hot-swapping the containers substrate
uses for given services. We'll do this by pre-resolving the image tags
to image ids. For now just adjust the defs to work with both the old and
new approaches.

While making this change, drop docker compose support. This reduces our
support burden and makes the project easier to understand.