feat: substrate builds faster and reloads services
Substrate now builds containers in parallel and will reload any services
whose definitions have changed. This is still experimental.

Podman-based builds also use much more parallelism now. Support for
docker is removed in this commit, to reduce the support vector.

Reusing resourcedirs is much faster now, as we now skip running a
resourcedir container if the target directory exists. Care must be taken
to avoid partially defined resourcedirs. Similar concerns exist if the
implementation of a resourcedir container changes as this will *not* be
detected or re-run.