feat: use container images as resourcedirs
Switch from resourcedirs being folders we manage to being container
images we mount. This simplifies a lot of things including our build
process, service spawning, even OS installation.

This change is a bit bigger than I had hoped it would be because
it requires some unfortunate workarounds. The most intricate of these
is avoiding building resourcedirs that we already have. To avoid
this extra overhead we would need a way to prevent `bakex build` from
sending layers that we already have.

Switch mounts to be an object in cue instead of a list. This makes it
much easier to have mounts be a derived value, since we don't need define
them all in a single list expression.

Clean up how we generate systemd units and presets. They are now a bit
more maintainable and easier to reason about.

Use rsync in reload instead of cp. This avoids changing files that
haven't changed and reduces the noise experienced by our autoreload
logic used during development.

Auto restart services when unit files update. This uses a few systemd
concepts and is a bit more complex than I'd like. But it seems to work
alright for now.

Overall this commit should improve the developer experience a fair bit.
We should not see spurious service restarts on reload and noop builds
should be much faster than they were.

Unfortunately the "first build" after this change will be slow, since
resourcedirs need to be built as container images and doing that is...
slow.

Future work will enable "focus mode", better caching, as well as
incremental layer loading.