fix: Decode Event.Response as any
When we switched resourcedirs to being container images, the mount
data structure switched from being a list to being an object. This
means we can't deserialize the ServiceSpawnResponse in past events.

So switch to loading it as any. This works for our needs and keeps
things both simple and flexible.