command-panel: use REFLECT for main page
Move the iframe message-passing to `/iframe` for
now and make `/` just use REFLECT on the URL to
look up commands.

We can have this automatically detected later if
the iframe responds to the message to export its
commands, but just getting this working for more
compatibility.

Also re-enables the tool-call suggestions.