load command-panel in bridge from command-panel service
This is a bit hacky, but testing this out. The
vite dev server is configured by SvelteKit with
only a limited set of directories to serve from.
Copying the files to the `static` directory is
hidden since override that with a mount from the
source directory.

The `__package__` directory also is not
re-generated by vite, but I didn't see a file
in the vite directories that had the expected
exports we'd need to load the <command-panel>
component.

So, this could use some more work, but wanted to
get something in that would load the latest
command-panel without manual copying.