feat: simplify provisioner, improve hairpin client
Implement the provisioner in a way that is better aligned with what
we've learned from building and using it. The provisioner essentially
delivers requests to a single backend for each URL path prefix it
encounters. There are circumstances where there is no valid backend,
so we introduce the idea of a ResolveError that can provide an uncached
http.Handler for that individual response (or responses, if many arrive
together and are coalesced).

Based on this insight, factor caching and blocking logic into a separate
handler and simplify our provisioner implementation to use it.

Also fix some functionality downstream of these changes.

Our hairpin http client needs to handle redirects in a way that's
similar to how the default http client handles them. So we import
corresponding logic from the go stdlib.

Switch new:instance to use a HEAD request and follow redirects before
returning the URL of the new instance.