feat: an interpreter for commands
Commands now have a simpler and more rigorous footing. They are based
on a data structure that supports recursive nesting and is easily
extensible.

This somewhat increases the complexity of the client logic needed to
execute a command, that logic is fairly simple and easy to port.

Also start replacing the name "command" and "def" with "msg" in a few
places. Future commits will continue this.