fix: `./remote mount NAME DIR` uses a system dir
Using a system dir for remote mounts makes a bit more sense than using
the source dir since the things we mount won't necessarily be source.

Using /mnt specifically also makes it easier to ensure the directory
actually exists when substrate starts, which is what the current
bug in main is due to.