add "events:try-reaction" command
I started testing this out as a simple way to
prototype reactions. The input is the same
as `events:write` (though right now limited to 1
reaction) which makes it easy to save by just
updating the command to write it when you're done.

But instead if runs and returns the results of the
reaction.

This made it fairly easy to iterate on the
reaction for combining transcription words that we
were looking at earlier.