try porting TS capreflectedmsg
This was an attempt to port the TypeScript implementation. It does execute now, but does not seem to fill the output.
Also, the TS version used an array in some places I had to replace with Fields using stringified indexes as the field names. It works with the existing Go implementation of `seq`, but I'm not sure that's really the intended structure here.